On branch master Your branch is up to date with 'origin/master'.
10 Juli 2023
- Fix cetak transaksi
- Menambahkan fungsi logout untuk akun petugas yang berhasil ditambahkan, dan login ulang untuk admin (Belum test)

Changes to be committed:
	modified:   Todo.txt
	modified:   lib/app/screens/home/admin/add_petugas/add_petugas_controller.dart
	modified:   lib/app/screens/home/petugas/request_pengambilan/request_pengambilan_controller.dart

+++ lib/app/screens/home/petugas/request_pengambilan/request_pengambilan_controller.dart
@@ -59,7 +59,7 @@ class RequestPengambilanController extends GetxController {
         //* Update Jumlah Poin Warga //
         await usersDb.doc(email).update({
           "poin": FieldValue.increment(
-              (jumlahOrganik * 150) + (jumlahAnorganik * 200))
+              (jumlahOrganik * 150) + (jumlahAnorganik * 200) as int)
         });
         //* Update Transaksi Sampah pada Collection transaksiSampah //
         await transaksiDb.doc(id).update({
@@ -96,7 +96,7 @@ class RequestPengambilanController extends GetxController {
         //* Update Jumlah Poin Warga //
         await usersDb.doc(email).update({
           "poin": FieldValue.increment(
-              (jumlahOrganik * 150) + (jumlahAnorganik * 200))
+              (jumlahOrganik * 150) + (jumlahAnorganik * 200) as int)
         });
         //* Update Transaksi Sampah pada Collection transaksiSampah //
         await transaksiDb.doc(id).update({
@@ -133,7 +133,7 @@ class RequestPengambilanController extends GetxController {
         //* Update Jumlah Poin Warga //
         await usersDb.doc(email).update({
           "poin": FieldValue.increment(
-              (jumlahOrganik * 150) + (jumlahAnorganik * 200))
+              (jumlahOrganik * 150) + (jumlahAnorganik * 200) as int)
         });
         //* Update Transaksi Sampah pada Collection transaksiSampah //
         await transaksiDb.doc(id).update({
@@ -170,7 +170,7 @@ class RequestPengambilanController extends GetxController {
         //* Update Jumlah Poin Warga //
         await usersDb.doc(email).update({
           "poin": FieldValue.increment(
-              (jumlahOrganik * 150) + (jumlahAnorganik * 300))
+              (jumlahOrganik * 150) + (jumlahAnorganik * 300) as int)
         });
         //* Update Transaksi Sampah pada Collection transaksiSampah //
         await transaksiDb.doc(id).update({
@@ -207,7 +207,7 @@ class RequestPengambilanController extends GetxController {
         //* Update Jumlah Poin Warga //
         await usersDb.doc(email).update({
           "poin": FieldValue.increment(
-              (jumlahOrganik * 150) + (jumlahAnorganik * 320))
+              (jumlahOrganik * 150) + (jumlahAnorganik * 320) as int)
         });
         //* Update Transaksi Sampah pada Collection transaksiSampah //
         await transaksiDb.doc(id).update({
@@ -244,7 +244,7 @@ class RequestPengambilanController extends GetxController {
         //* Update Jumlah Poin Warga //
         await usersDb.doc(email).update({
           "poin": FieldValue.increment(
-              (jumlahOrganik * 150) + (jumlahAnorganik * 220))
+              (jumlahOrganik * 150) + (jumlahAnorganik * 220) as int)
         });
         //* Update Transaksi Sampah pada Collection transaksiSampah //
         await transaksiDb.doc(id).update({
